+++ CMPS278Backend/ModelsDTO/BaseModelDTOs/BaseReviewModelDTO.cs
@@ -6,7 +6,7 @@ public abstract class BaseReviewModelDTO
     public string? UserName  { get; set; }
     public string? UserImage { get; set; }
     public string? Date      { get; set; }
-    public float?  Score     { get; set; }
+    public double?  Score     { get; set; }
     public string? ScoreText { get; set; }
     public string? Url       { get; set; }
     public string? Title     { get; set; }
@@ -14,5 +14,5 @@ public abstract class BaseReviewModelDTO
     public string? ReplyDate { get; set; }
     public string? ReplyText { get; set; }
     public string? Version   { get; set; }
-    public int?    ThumbsUp  { get; set; }
+    public long?    ThumbsUp  { get; set; }
 }
